feature crud floor

+++ src/main/kotlin/com/dataner/domain/floor/services/FloorServiceImpl.kt
@@ -8,7 +8,6 @@ class FloorServiceImpl(
     private val floorRepository: FloorRepository
 ) : FloorService {
     override fun createFloor(floor: Floor) {
-
         if (!floorRepository.checkFloorNumber(floor))
             throw Exception()
 
@@ -24,6 +23,9 @@ class FloorServiceImpl(
         floorRepository.allBuildingFloors(buildingId = buildingId)
 
     override fun deleteFloor(floorId: Int) {
+        if (floorRepository.checkFloor(floorId))
+            throw Exception()
+
         floorRepository.deleteFloor(floorId = floorId)
     }
 }
